[Refactor] 코드 정리 및 쿼리 파라미터에 따라 지시문 변경
